move to comparison to new function; return as json
